Refactor switcher into main
