--- benchmark/include/switcher.h
@@ -1,5 +1,0 @@
-#pragma once
-#include <atomic>
-#include "scenario.h"
-
-void startSwitcher(const ScenarioFile& config, std::atomic<bool>& running);
